+++ include/engine/datafacade/contiguous_internalmem_datafacade.hpp
@@ -14,6 +14,7 @@
 #include "extractor/datasources.hpp"
 #include "extractor/edge_based_node.hpp"
 #include "extractor/intersection_bearings_container.hpp"
+#include "extractor/maneuver_override.hpp"
 #include "extractor/node_data_container.hpp"
 #include "extractor/packed_osm_ids.hpp"
 #include "extractor/profile_properties.hpp"
@@ -204,6 +205,9 @@ class ContiguousInternalMemoryDataFacadeBase : public BaseDataFacade
     util::vector_view<std::size_t> m_datasource_name_offsets;
     util::vector_view<std::size_t> m_datasource_name_lengths;
     util::vector_view<util::guidance::LaneTupleIdPair> m_lane_tupel_id_pairs;
+
+    util::vector_view<extractor::StorageManeuverOverride> m_maneuver_overrides;
+    util::vector_view<NodeID> m_maneuver_override_node_sequences;
 
     std::unique_ptr<SharedRTree> m_static_rtree;
     std::unique_ptr<SharedGeospatialQuery> m_geospatial_query;
@@ -499,6 +503,21 @@ class ContiguousInternalMemoryDataFacadeBase : public BaseDataFacade
         m_entry_class_table = std::move(entry_class_table);
     }
 
+    void InitializeManeuverOverridePointers(storage::DataLayout &data_layout, char *memory_block)
+    {
+        auto maneuver_overrides_ptr = data_layout.GetBlockPtr<extractor::StorageManeuverOverride>(
+            memory_block, storage::DataLayout::MANEUVER_OVERRIDES);
+        m_maneuver_overrides = util::vector_view<extractor::StorageManeuverOverride>(
+            maneuver_overrides_ptr,
+            data_layout.num_entries[storage::DataLayout::MANEUVER_OVERRIDES]);
+
+        auto maneuver_override_node_sequences_ptr = data_layout.GetBlockPtr<NodeID>(
+            memory_block, storage::DataLayout::MANEUVER_OVERRIDE_NODE_SEQUENCES);
+        m_maneuver_override_node_sequences = util::vector_view<NodeID>(
+            maneuver_override_node_sequences_ptr,
+            data_layout.num_entries[storage::DataLayout::MANEUVER_OVERRIDE_NODE_SEQUENCES]);
+    }
+
     void InitializeInternalPointers(storage::DataLayout &data_layout,
                                     char *memory_block,
                                     const std::size_t exclude_index)
@@ -515,6 +534,7 @@ class ContiguousInternalMemoryDataFacadeBase : public BaseDataFacade
         InitializeProfilePropertiesPointer(data_layout, memory_block, exclude_index);
         InitializeRTreePointers(data_layout, memory_block);
         InitializeIntersectionClassPointers(data_layout, memory_block);
+        InitializeManeuverOverridePointers(data_layout, memory_block);
     }
 
   public:
@@ -889,6 +909,39 @@ class ContiguousInternalMemoryDataFacadeBase : public BaseDataFacade
     bool IsSegregated(const NodeID id) const override final
     {
         return edge_based_node_data.IsSegregated(id);
+    }
+
+    std::vector<extractor::ManeuverOverride>
+    GetOverridesThatStartAt(const NodeID edge_based_node_id) const override final
+    {
+        std::vector<extractor::ManeuverOverride> results;
+
+        // heterogeneous comparison:
+        struct Comp
+        {
+            bool operator()(const extractor::StorageManeuverOverride &s, NodeID i) const
+            {
+                return s.start_node < i;
+            }
+            bool operator()(NodeID i, const extractor::StorageManeuverOverride &s) const
+            {
+                return i < s.start_node;
+            }
+        };
+
+        auto found_range = std::equal_range(
+            m_maneuver_overrides.begin(), m_maneuver_overrides.end(), edge_based_node_id, Comp{});
+
+        std::for_each(found_range.first, found_range.second, [&](const auto & override) {
+            std::vector<NodeID> sequence(
+                m_maneuver_override_node_sequences.begin() + override.node_sequence_offset_begin,
+                m_maneuver_override_node_sequences.begin() + override.node_sequence_offset_end);
+            results.push_back(extractor::ManeuverOverride{std::move(sequence),
+                                                          override.instruction_node,
+                                                          override.override_type,
+                                                          override.direction});
+        });
+        return results;
     }
 };