+++ include/engine/guidance/route_step.hpp
@@ -59,6 +59,7 @@ inline IntermediateIntersection getInvalidIntersection()
 
 struct RouteStep
 {
+    NodeID from_id;
     unsigned name_id;
     bool is_segregated;
     std::string name;