+++ include/storage/shared_datatype.hpp
@@ -88,7 +88,9 @@ const constexpr char *block_id_to_name[] = {"NAME_CHAR_DATA",
                                             "MLD_CELL_LEVEL_OFFSETS",
                                             "MLD_GRAPH_NODE_LIST",
                                             "MLD_GRAPH_EDGE_LIST",
-                                            "MLD_GRAPH_NODE_TO_OFFSET"};
+                                            "MLD_GRAPH_NODE_TO_OFFSET",
+                                            "MANEUVER_OVERRIDES",
+                                            "MANEUVER_OVERRIDE_NODE_SEQUENCES"};
 
 struct DataLayout
 {
@@ -165,6 +167,8 @@ struct DataLayout
         MLD_GRAPH_NODE_LIST,
         MLD_GRAPH_EDGE_LIST,
         MLD_GRAPH_NODE_TO_OFFSET,
+        MANEUVER_OVERRIDES,
+        MANEUVER_OVERRIDE_NODE_SEQUENCES,
         NUM_BLOCKS
     };