+++ src/extractor/node_based_graph_factory.cpp
@@ -19,10 +19,14 @@ NodeBasedGraphFactory::NodeBasedGraphFactory(
     const boost::filesystem::path &input_file,
     ScriptingEnvironment &scripting_environment,
     std::vector<TurnRestriction> &turn_restrictions,
-    std::vector<ConditionalTurnRestriction> &conditional_turn_restrictions)
+    std::vector<ConditionalTurnRestriction> &conditional_turn_restrictions,
+    std::vector<UnresolvedManeuverOverride> &maneuver_overrides)
 {
     LoadDataFromFile(input_file);
-    Compress(scripting_environment, turn_restrictions, conditional_turn_restrictions);
+    Compress(scripting_environment,
+             turn_restrictions,
+             conditional_turn_restrictions,
+             maneuver_overrides);
     CompressGeometry();
     CompressAnnotationData();
 }
@@ -84,7 +88,8 @@ void NodeBasedGraphFactory::LoadDataFromFile(const boost::filesystem::path &inpu
 void NodeBasedGraphFactory::Compress(
     ScriptingEnvironment &scripting_environment,
     std::vector<TurnRestriction> &turn_restrictions,
-    std::vector<ConditionalTurnRestriction> &conditional_turn_restrictions)
+    std::vector<ConditionalTurnRestriction> &conditional_turn_restrictions,
+    std::vector<UnresolvedManeuverOverride> &maneuver_overrides)
 {
     GraphCompressor graph_compressor;
     graph_compressor.Compress(barriers,
@@ -92,6 +97,7 @@ void NodeBasedGraphFactory::Compress(
                               scripting_environment,
                               turn_restrictions,
                               conditional_turn_restrictions,
+                              maneuver_overrides,
                               compressed_output_graph,
                               annotation_data,
                               compressed_edge_container);